+++ SteamPS/Private/Get-SteamPath.ps1
@@ -6,12 +6,12 @@
     process {
         $SteamCMDPath = $env:Path.Split(';') | Where-Object -FilterScript { $_ -like "*SteamCMD*" }
         if ($null -ne $SteamCMDPath) {
-            $ObjectProperties = [ordered]@{
-                'Path'       = $SteamCMDPath;
-                'Executable' = "$($SteamCMDPath)\steamcmd.exe";
+            [PSCustomObject]@{
+                'Path'       = $SteamCMDPath
+                'Executable' = "$SteamCMDPath\steamcmd.exe"
             }
-
-            New-Object –TypeName PSObject –Property $ObjectProperties
-        } # if
+        } else {
+            Write-Verbose -Message 'SteamCMD where not found on the environment path.'
+        }
     } # Process
 } # Cmdlet