+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### Changed
+
+- Improving structure / format of the code.
+  - Remove the subexpression where they are not needed.
+
 ## [3.0.0] - 19/05-2020
 
 ### Fixed